refactor: slightly refactored schemas and models
